MDEV-35506 commit policy of one-phase-commit even at errored-out binlogging leads to assert
Currently execution of commit in one phase proceeds to commit by
engines when binlog_commit() does not succeed.

There are two issues with that:

1. absence of binlog_rollback() or lower-level
`binlog_cache_data::reset()` along the following execution of the
failing statement eventually will raise an assert on non-empty binlog
cache;
2. engines, including ones that are rollback capable, commit in this
particular error situation.

Both effects are demonstrated with an mtr test results file, recorded
on *BASE* of the commit.

BASE has to include MDEV-35207 commit because the test is written
with CREATE-TABLE-SELECTs.

Todo: explore why Aria seems to be unable to rollback the current
      statement by its rollback().